+++ src/PopoverNavigation.js
@@ -1,8 +1,8 @@
-import Popover, { Rect, PLACEMENT_OPTIONS } from './Popover'
+import Popover from './Popover'
+import { popoverTransitionConfig, Rect, PLACEMENT_OPTIONS, isIOS } from './Utility'
 import React, { Component } from 'react'
-import { View, Platform, BackHandler, Dimensions, Animated } from 'react-native'
+import { View, BackHandler, Dimensions, Animated } from 'react-native'
 import PropTypes from 'prop-types'
-import { StackNavigator } from 'react-navigation'
 
 var {height: SCREEN_HEIGHT, width: SCREEN_WIDTH} = Dimensions.get('window');
 const defaultDisplayArea = new Rect(10, 10, SCREEN_WIDTH-20, SCREEN_HEIGHT-20);
@@ -10,68 +10,19 @@ const defaultDisplayArea = new Rect(10, 10, SCREEN_WIDTH-20, SCREEN_HEIGHT-20);
 let popoverDisplayAreaChangeListeners = [];
 let displayAreaStore = null;
 
-let shouldShowInPopover = () => true;
-
-function addPopoverInstance(instance) {
-  popoverDisplayAreaChangeListeners.push(instance);
-  return () => popoverDisplayAreaChangeListeners.splice(popoverDisplayAreaChangeListeners.indexOf(instance), 1);
-}
-
-// Transition config needed on tablets for popover to work
-export let popoverTransitionConfig = () => ({
-  transitionSpec: {
-    duration: 1,
-    timing: Animated.timing,
-  },
-  screenInterpolator: sceneProps => {
-    const { position, scene } = sceneProps
-    const { index } = scene
-
-    const translateY = position.interpolate({
-      inputRange: [index - 1, index, index + 1],
-      outputRange: [0, 0, 0],
-    })
-
-    const opacity = position.interpolate({
-      inputRange: [index - 1, index, index + 1],
-      outputRange: [0, 1, 1],
-    })
-
-    return { opacity, transform: [{ translateY }] }
-  },
-})
+export let shouldShowInPopover = () => true;
 
 export let withPopoverNavigationRootWrapper = Comp => props => <View 
   style={{position: 'absolute', left: 0, right: 0, top: 0, bottom: 0}}
   onLayout={evt => {
     let popoverDispayArea = {
       x: 10,
-      y: Platform.OS === 'ios' ? 20 : 10,
+      y: isIOS() ? 20 : 10,
       width: evt.nativeEvent.layout.width - 20,
-      height: evt.nativeEvent.layout.height - (Platform.OS === 'ios' ? 30 : 20)
+      height: evt.nativeEvent.layout.height - (isIOS() ? 30 : 20)
     }
     PopoverNavigation.setDisplayArea(popoverDispayArea)
   }}><Comp {...props} /></View>
-
-export let withPopoverNavigation = (Comp, popoverOptions) => props => <PopoverNavigation {...popoverOptions}><Comp {...props} /></PopoverNavigation>;
-
-export function PopoverStackNavigator(RouteConfigs, StackNavigatorConfig) {
-  let routeKeys = Object.keys(RouteConfigs);
-  let firstRouteKey = routeKeys.splice(0, 1)[0];
-  let newRouteConfigs = {};
-  newRouteConfigs[firstRouteKey] = RouteConfigs[firstRouteKey];
-  routeKeys.forEach(route => newRouteConfigs[route] = Object.assign({}, RouteConfigs[route], { screen: withPopoverNavigation(RouteConfigs[route].screen, RouteConfigs[route].popoverOptions) }));
-
-  let nonPopoverStack = StackNavigator(newRouteConfigs, StackNavigatorConfig);
-  const popoverStackConfig = Object.assign({}, StackNavigatorConfig, {transitionConfig: popoverTransitionConfig, headerMode: 'screen', cardStyle: {backgroundColor: 'transparent'}});
-  let popoverStack = StackNavigator(newRouteConfigs, popoverStackConfig);
-
-  return props => {
-    let Stack = nonPopoverStack;
-    if (shouldShowInPopover()) Stack = popoverStack;
-    return <Stack screenProps={props.screenProps} ref={props.navigatorRef} />;
-  }
-}
 
 export default class PopoverNavigation extends Component {
   static navigationOptions = {}
@@ -92,6 +43,11 @@ export default class PopoverNavigation extends Component {
     }
   }
 
+  static addPopoverInstance(instance) {
+    popoverDisplayAreaChangeListeners.push(instance);
+    return () => popoverDisplayAreaChangeListeners.splice(popoverDisplayAreaChangeListeners.indexOf(instance), 1);
+  }
+
   goBack() {
     if (shouldShowInPopover())
       this.setState({visible: false});
@@ -101,7 +57,7 @@ export default class PopoverNavigation extends Component {
 
   componentDidMount() {
     BackHandler.addEventListener('hardwareBackPress', this.backButtonPressed);
-    this.removeDisplayAreaChangeListener = addPopoverInstance(this);
+    this.removeDisplayAreaChangeListener = PopoverNavigation.addPopoverInstance(this);
     this.saveStashRect()
   }
