+++ src/index.js
@@ -1,11 +1,13 @@
-import Popover, { Rect, PLACEMENT_OPTIONS } from './Popover';
-import PopoverNavigation, { popoverTransitionConfig, PopoverStackNavigator, withPopoverNavigation, withPopoverNavigationRootWrapper } from './PopoverNavigation';
+import Popover from './Popover'
+import PopoverNavigation, { withPopoverNavigationRootWrapper } from './PopoverNavigation'
+import PopoverStackNavigator, { withPopoverNavigation } from './PopoverStackNavigator'
+import { Rect, PLACEMENT_OPTIONS, popoverTransitionConfig } from './Utility'
 
-export default Popover;
-export { PopoverNavigation as PopoverNavigation };
-export { Rect as Rect };
-export { PLACEMENT_OPTIONS as PLACEMENT_OPTIONS };
-export { popoverTransitionConfig as popoverTransitionConfig };
-export { PopoverStackNavigator as PopoverStackNavigator };
-export { withPopoverNavigation as withPopoverNavigation };
-export { withPopoverNavigationRootWrapper as withPopoverNavigationRootWrapper };
+export default Popover
+export { PopoverNavigation as PopoverNavigation }
+export { Rect as Rect }
+export { PLACEMENT_OPTIONS as PLACEMENT_OPTIONS }
+export { popoverTransitionConfig as popoverTransitionConfig }
+export { PopoverStackNavigator as PopoverStackNavigator }
+export { withPopoverNavigation as withPopoverNavigation }
+export { withPopoverNavigationRootWrapper as withPopoverNavigationRootWrapper }
